requirements.txt
aiohttp==3.9.1
python-dotenv==1.0.0
openai==1.3.7
deep-translator==1.11.4
feedparser==6.0.10
pyyaml==6.0.1
beautifulsoup4==4.12.2
requests>=2.31.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
